messaging queue/requirements.txt
psycopg2-binary==2.9.9
pika==1.3.2
python-dotenv==1.0.0
jinja2==3.1.2
pytest==7.4.3
pytest-asyncio==0.21.1
redis==5.0.1
pytz==2023.3